Restructure tests to separate deployments
