+++ .circleci/config.yml
@@ -171,7 +171,7 @@ jobs:
 
       - run:
           name: Run anvil localhost:8545
-          command: anvil --fork-url "<< parameters.provider-url >>"
+          command: anvil --auto-impersonate --chain-id "<< parameters.chainId >>" --fork-url "<< parameters.provider-url >>"
           background: true
 
       - run:
@@ -184,7 +184,7 @@ jobs:
 
       - run:
           working_directory: /tmp/app/liquidity/cypress
-          command: cypress run --e2e --browser chrome
+          command: cypress run --e2e --browser chrome --config specPattern="./cypress/e2e/<< parameters.chainId >>-<< parameters.preset >>/*.e2e.js"
 
       - store_test_results:
           path: 'liquidity/cypress/cypress/reports'
@@ -281,15 +281,24 @@ workflows:
           chainId: 8453
           preset: andromeda
           provider-url: https://base-mainnet.infura.io/v3/$INFURA_KEY
+      - liquidity-e2e:
+          name: liquidity-e2e-arbitrum-mainnet
+          chainId: 42161
+          preset: main
+          provider-url: https://arbitrum-mainnet.infura.io/v3/$INFURA_KEY
       - combine-coverage:
-          requires: [tests, liquidity-cy, liquidity-e2e-base-mainnet]
-
+          requires:
+            - tests
+            - liquidity-cy
+            - liquidity-e2e-base-mainnet
+            - liquidity-e2e-arbitrum-mainnet
       - liquidity-ipfs:
           requires:
             - checks
             - tests
             - liquidity-cy
             - liquidity-e2e-base-mainnet
+            - liquidity-e2e-arbitrum-mainnet
           filters:
             branches:
               only:
