+++ liquidity/cypress/cypress.config.js
@@ -45,7 +45,6 @@ module.exports = defineConfig({
         ...require('./cypress/tasks/setConfig'),
         ...require('./cypress/tasks/getSUSDC'),
         ...require('./cypress/tasks/getUSDC'),
-        ...require('./cypress/tasks/isBase'),
       });
 
       return config;
