+++ liquidity/cypress/cypress/e2e/8453-andromeda/Create Account.e2e.js
@@ -1,0 +1,24 @@
+it('Create Account', () => {
+  cy.connectWallet().then((wallet) => {
+    cy.task('setEthBalance', { address: wallet.address, balance: 2 });
+  });
+
+  cy.viewport(1200, 900);
+  cy.visit('/');
+
+  cy.get('[data-cy="account-menu-button"]').click();
+  cy.contains('[data-cy="create-new-account-menu-item"]', 'Create Account').should('exist');
+
+  // Create account
+  cy.get('[data-cy="header-wallet-address-button"]').click();
+  cy.get('[data-cy="create-new-account-menu-item"]').click();
+  cy.get('[data-cy="header-account-list"]').children().should('have.length', 1);
+
+  // The account id in the url should be the same as in the header
+  cy.url().then((url) => {
+    const [baseUrl, fragment] = url.split('/#/');
+    const params = new URL(`${baseUrl}${fragment}`).searchParams;
+    const accountId = params.get('accountId');
+    cy.get(`[data-cy="account-${accountId}"]`).should('exist');
+  });
+});
